+++ src/apps/toolbar/modules/item/app.ts
@@ -1,12 +1,14 @@
 import { invoke } from '@tauri-apps/api/core';
 import { evaluate } from 'mathjs';
 
+/** @deprecated remove on v2 */
 export enum Actions {
   Open = 'open',
   CopyToClipboard = 'copy-to-clipboard',
   SwitchWorkspace = 'switch-workspace',
 }
 
+/** @deprecated remove on v2 */
 export function performClick(onClick: string | null, scope: any) {
   if (!onClick) {
     return;
@@ -34,5 +36,55 @@ export function performClick(onClick: string | null, scope: any) {
       if (argument) {
         invoke('switch_workspace', { idx: evaluate(argument, scope) });
       }
+  }
+}
+
+export class Scope {
+  scope: Map<string, any>;
+
+  constructor() {
+    this.scope = new Map();
+  }
+
+  get(key: string) {
+    return this.scope.get(key);
+  }
+
+  set(key: string, value: any) {
+    return this.scope.set(key, value);
+  }
+
+  has(key: string) {
+    return this.scope.has(key);
+  }
+
+  keys(): string[] | IterableIterator<string> {
+    return this.scope.keys();
+  }
+
+  loadInvokeActions() {
+    for (const [key, value] of Object.entries(ActionsScope)) {
+      this.set(key, value);
+    }
+  }
+}
+
+const ActionsScope = {
+  open(path: string) {
+    invoke('open_file', { path }).catch(console.error);
+  },
+  run(program: string, ...args: string[]) {
+    invoke('run', { program, args }).catch(console.error);
+  },
+  copyClipboard(text: string) {
+    navigator.clipboard.writeText(text);
+  },
+};
+
+export function safeEval(expression: string, scope: Scope) {
+  try {
+    evaluate(expression, scope);
+  } catch (error) {
+    console.error(error);
   }
 }