+++ src/apps/toolbar/modules/item/infra.tsx
@@ -1,24 +1,21 @@
-import {
-  exposedIcons,
-  exposedIconsRegex,
-  Icon,
-  IconName,
-  isValidIconName,
-} from '../../../shared/components/Icon';
+import { exposedIcons, Icon, IconName } from '../../../shared/components/Icon';
 import { ToolbarModule } from '../../../shared/schemas/Placeholders';
 import { cx } from '../../../shared/styles';
+import { convertFileSrc, invoke } from '@tauri-apps/api/core';
 import { Tooltip } from 'antd';
 import { Reorder } from 'framer-motion';
 import { cloneDeep } from 'lodash';
 import { evaluate } from 'mathjs';
-import React, { useEffect, useRef } from 'react';
+import React, { PropsWithChildren, useEffect, useRef } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useSelector } from 'react-redux';
 
-import { Selectors } from '../shared/store/app';
-import { performClick } from './app';
+import { LAZY_CONSTANTS } from '../shared/utils/infra';
 
-interface Props {
+import { Selectors } from '../shared/store/app';
+import { performClick, safeEval, Scope } from './app';
+
+interface Props extends PropsWithChildren {
   module: ToolbarModule;
   extraVars?: Record<string, any>;
   active?: boolean;
@@ -27,59 +24,115 @@ interface Props {
   onKeydown?: (e: React.KeyboardEvent) => void;
 }
 
-class Scope {
-  scope: Map<string, any>;
+interface StringToElementProps {
+  text: string;
+}
 
-  constructor() {
-    this.scope = new Map();
+interface StringToElementState {
+  exe_icon_path: string;
+}
+
+class StringToElement extends React.Component<StringToElementProps, StringToElementState> {
+  static splitter = /:([^:]+):/;
+
+  static imgPrefix = 'IMG:';
+  static iconPrefix = 'ICON:';
+  static exePrefix = 'EXE:';
+
+  static imgFromUrl(url: string, size = 16) {
+    return `[IMG:${size}px:${url}]`;
   }
 
-  get(key: string) {
-    return this.scope.get(key);
+  static imgFromPath(path: string, size = 16) {
+    return StringToElement.imgFromUrl(convertFileSrc(path), size);
   }
 
-  set(key: string, value: any) {
-    return this.scope.set(key, value);
+  static imgFromExe(exe_path: string, size = 16) {
+    return `[EXE:${size}px:${exe_path}]`;
   }
 
-  has(key: string) {
-    return this.scope.has(key);
+  constructor(props: StringToElementProps) {
+    super(props);
+    this.state = { exe_icon_path: LAZY_CONSTANTS.MISSING_ICON_PATH };
   }
 
-  keys(): string[] | IterableIterator<string> {
-    return this.scope.keys();
+  isImg() {
+    return this.props.text.startsWith(StringToElement.imgPrefix);
+  }
+
+  isIcon() {
+    return this.props.text.startsWith(StringToElement.iconPrefix);
+  }
+
+  isExe() {
+    return this.props.text.startsWith(StringToElement.exePrefix);
+  }
+
+  setExeIcon(exe_path: string | null) {
+    this.setState({ exe_icon_path: exe_path || LAZY_CONSTANTS.MISSING_ICON_PATH });
+  }
+
+  componentDidMount() {
+    if (this.isExe()) {
+      const [_, _size, path] = this.props.text.split(StringToElement.splitter);
+      if (path) {
+        invoke<string | null>('get_icon', { path })
+          .then(this.setExeIcon.bind(this))
+          .catch(console.error);
+      }
+    }
+  }
+
+  render() {
+    if (this.isExe()) {
+      const [_, width] = this.props.text.split(StringToElement.splitter);
+      return <img src={convertFileSrc(this.state.exe_icon_path)} style={{ width }} />;
+    }
+
+    if (this.isImg()) {
+      const [_, width, url] = this.props.text.split(StringToElement.splitter);
+      return <img src={url} style={{ width }} />;
+    }
+
+    if (this.isIcon()) {
+      const [_, iconName, size] = this.props.text.split(':');
+      return (
+        <Icon
+          iconName={iconName as IconName}
+          propsIcon={{ size: size ? size + 'px' : undefined }}
+        />
+      );
+    }
+
+    return <span>{this.props.text}</span>;
   }
 }
 
 export function ElementsFromEvaluated(content: any) {
+  let text: string = content;
+
   if (typeof content !== 'string') {
-    content = JSON.stringify(content);
+    text = JSON.stringify(content);
   }
 
-  const parts: string[] = content.split(exposedIconsRegex);
-  const result: React.ReactNode[] = [];
-
-  parts.forEach((part: string, index: number) => {
-    if (isValidIconName(part)) {
-      const [iconName, size] = part.split(':') as [IconName, string?];
-      result.push(
-        <Icon
-          key={index}
-          iconName={iconName}
-          propsIcon={{ size: size ? size + 'px' : undefined }}
-        />,
-      );
-    } else if (part) {
-      result.push(<span key={index}>{part}</span>);
-    }
+  const parts: string[] = text.split(/\[|\]/g).filter((part: string) => part);
+  const result: React.ReactNode[] = parts.map((part: string, index: number) => {
+    return <StringToElement text={part} key={index} />;
   });
 
   return result;
 }
 
 export function Item(props: Props) {
-  const { extraVars, module, active, onClick: onClickProp, onKeydown: onKeydownProp } = props;
-  const { template, tooltip, onClick, style, id } = module;
+  const {
+    extraVars,
+    module,
+    active,
+    onClick: onClickProp,
+    onKeydown: onKeydownProp,
+    children,
+  } = props;
+  const { template, tooltip, onClick: oldOnClick, onClickV2, style, id } = module;
 
   const [mounted, setMounted] = React.useState(false);
   const env = useSelector(Selectors.env);
@@ -90,14 +143,23 @@ export function Item(props: Props) {
 
   const { t } = useTranslation();
   const scope = useRef(new Scope());
-  scope.current.set('t', t);
 
   useEffect(() => {
+    scope.current.loadInvokeActions();
+
     scope.current.set('icon', cloneDeep(exposedIcons));
     scope.current.set('env', cloneDeep(env));
+    scope.current.set('imgFromUrl', StringToElement.imgFromUrl);
+    scope.current.set('imgFromPath', StringToElement.imgFromPath);
+    scope.current.set('imgFromExe', StringToElement.imgFromExe);
     setMounted(true);
   }, []);
 
+  if (!mounted) {
+    return null;
+  }
+
+  scope.current.set('t', t);
   scope.current.set('window', { ...window });
   if (extraVars) {
     Object.keys(extraVars).forEach((key) => {
@@ -105,12 +167,8 @@ export function Item(props: Props) {
     });
   }
 
-  if (!mounted) {
-    return null;
-  }
-
   const elements = ElementsFromEvaluated(evaluate(template, scope.current));
-  if (!elements.length) {
+  if (!elements.length && !children) {
     return null;
   }
 
@@ -125,19 +183,24 @@ export function Item(props: Props) {
         id={id}
         style={style}
         className={cx('ft-bar-item', {
-          'ft-bar-item-clickable': onClick || onClickProp,
+          'ft-bar-item-clickable': oldOnClick || onClickProp || onClickV2,
           'ft-bar-item-active': active,
         })}
         onKeyDown={onKeydownProp}
         onClick={(e) => {
           onClickProp?.(e);
-          performClick(onClick, scope.current);
+
+          if (onClickV2) {
+            safeEval(onClickV2, scope.current);
+          }
+
+          performClick(oldOnClick, scope.current);
         }}
         value={module}
         as="div"
         transition={{ duration: 0.15 }}
       >
-        <div className="ft-bar-item-content">{elements}</div>
+        <div className="ft-bar-item-content">{children || elements}</div>
       </Reorder.Item>
     </Tooltip>
   );